Version 0.8.7
* v0.7.8 update
  1. Supporting Wordpress REST API with OAuth 1.0a.
  2. Supporting Wordpress REST API with JWT Authentication.
  3. Fix few minor issues.

+++ WooCommerce/v3/Customer.cs
@@ -17,7 +17,7 @@ namespace WooCommerceNET.WooCommerce.v3
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? id { get; set; }
+        public long? id { get; set; }
 
         /// <summary>
         /// The date the customer was created, in the site’s timezone.
